add master user check.

+++ WebAPI/Controllers/PaymentGatewayController.cs
@@ -102,6 +102,17 @@ namespace WebAPI.Controllers
 
             try
             {
+                string userID = KS.GetFromRequest().UserId;
+
+                // get domain       
+                var domain = ClientsManager.DomainsClient().GetDomainByUser(groupId, userID);
+                
+                // check if the user performing the action is domain master
+                if (domain.MasterUsers.Where(u => u.Id == userID).FirstOrDefault() == null)
+                {
+                    throw new ForbiddenException();
+                }
+
                 // call client
                 response = ClientsManager.BillingClient().SetHouseholdPaymentGateway(groupId, payment_gateway_id, KS.GetFromRequest().UserId, household_id);
             }
@@ -118,8 +129,9 @@ namespace WebAPI.Controllers
         /// </summary>
         /// <remarks>
         /// Possible status codes:       
-        /// user does not exist = 2000, user not in domain = 1005, user with no domain = 2024, user suspended = 2001, domain not exists = 1006
-        /// payment gateway not exist = 6008, household not set to payment gateway = 6027
+        ///   User Does Not Exist = 2000, User Not In Domain = 1005, User With No Domain = 2024, User Suspended = 2001, Domain Not Exists = 1006
+        /// Payment Gateway Identifier is Missing = 6005, Payment gateway not exist = 6008, Household not set to payment gateway = 6027
+        /// PaymentGatewaySelectionIsDisabled = 6028, ServiceForbidden = 500004
         /// </remarks>
         /// <param name="payment_gateway_id">Payment Gateway Identifier</param>
         /// <param name="household_id">Household Identifier</param>
@@ -132,6 +144,17 @@ namespace WebAPI.Controllers
 
             try
             {
+                string userID = KS.GetFromRequest().UserId;
+
+                // get domain       
+                var domain = ClientsManager.DomainsClient().GetDomainByUser(groupId, userID);
+                
+                // check if the user performing the action is domain master
+                if (domain.MasterUsers.Where(u => u.Id == userID).FirstOrDefault() == null)
+                {
+                    throw new ForbiddenException();
+                }
+
                 // call client
                 response = ClientsManager.BillingClient().DeleteHouseholdPaymentGateway(groupId, payment_gateway_id, KS.GetFromRequest().UserId, household_id);
             }
